requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
google-api-core==1.20.0
google-auth==1.16.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.9
joblib==0.15.1
nltk==3.5
numpy==1.18.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
regex==2020.6.8
requests==2.23.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.46.1
urllib3==1.25.9
pyarrow==0.17.1